Clean up `NodeId` kind fetching
